+++ src/main/java/org/example/currency_exchanger/dto/ExchangeRateDto.java
@@ -1,4 +1,11 @@
 package org.example.currency_exchanger.dto;
 
-public record ExchangeRateDto(Long id, CurrencyDto baseCurrency, CurrencyDto targetCurrency, Double rate) {
+import java.math.BigDecimal;
+
+public record ExchangeRateDto(
+        Long id,
+        CurrencyDto baseCurrency,
+        CurrencyDto targetCurrency,
+        BigDecimal rate
+) {
 }
